Fixup Award points as float and fix type hinting.

+++ database/migrations/2025_08_25_040642_create_awards_table.php
@@ -27,7 +27,7 @@ return new class extends Migration
                 $table->string('image')->nullable();
                 $table->string('branch')->nullable();
                 $table->string('star_nation')->nullable();
-                $table->integer('points');
+                $table->float('points');
             });
         } else {
             echo "Table 'awards' already exists. Skipping migration.\n";
